requirements.txt
aiosqlite==0.21.0
fastapi==0.111.0
google-api-python-client==2.147.0
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
uvicorn==0.30.6
